Revert "Moved ExternalRandomNumberGeneratorService to GeneratorInterface"
