+++ FWCore/Services/BuildFile.xml
@@ -1,3 +1,4 @@
+<use name="SimDataFormats/RandomEngine"/>
 <use name="FWCore/AbstractServices"/>
 <use name="FWCore/MessageLogger"/>
 <use name="FWCore/ParameterSet"/>
@@ -7,6 +8,7 @@
 <use name="tinyxml2"/>
 <use name="rootcore"/>
 <use name="roothistmatrix"/>
+<use name="clhep"/>
 <export>
   <lib name="1"/>
 </export>
