+++ FWCore/Services/test/test_catch2_ExternalRandomNumberGeneratorService.cc
@@ -1,4 +1,4 @@
-#include "GeneratorInterface/Core/ExternalRandomNumberGeneratorService.h"
+#include "FWCore/Services/interface/ExternalRandomNumberGeneratorService.h"
 #include "FWCore/Utilities/interface/StreamID.h"
 
 #include "CLHEP/Random/JamesRandom.h"
