+++ GeneratorInterface/Core/bin/BuildFile.xml
@@ -3,14 +3,6 @@
   <use name="boost_program_options"/>
   <use name="FWCore/TestProcessor"/>
   <use name="FWCore/SharedMemory"/>
-  <use name="GeneratorInterface/Core"/>
+  <use name="FWCore/Services"/>
   <use name="SimDataFormats/GeneratorProducts"/>
-</bin>
-
-<bin name="cmsTestInterProcessRandom" file="interprocess_random.cc">
-  <use name="boost"/>
-  <use name="boost_program_options"/>
-  <use name="FWCore/TestProcessor"/>
-  <use name="FWCore/SharedMemory"/>
-  <use name="GeneratorInterface/Core"/>
 </bin>
