+++ GeneratorInterface/Core/plugins/BuildFile.xml
@@ -1,13 +1,11 @@
-<environment>
-  <library name="GeneratorInterfaceCore_plugins" file="*Gen*.cc">
-    <use name="FWCore/AbstractServices"/>
-    <use name="FWCore/Framework"/>
-    <use name="FWCore/ParameterSet"/>
-    <use name="FWCore/MessageLogger"/>
-    <use name="FWCore/Utilities"/>
-    <use name="SimDataFormats/GeneratorProducts"/>
-    <use name="FWCore/SharedMemory"/>
-    <use name="clhep"/>
-    <flags EDM_PLUGIN="1"/>
-  </library>
-</environment>
+<use name="FWCore/AbstractServices"/>
+<use name="FWCore/Framework"/>
+<use name="FWCore/ParameterSet"/>
+<use name="FWCore/MessageLogger"/>
+<use name="FWCore/Utilities"/>
+<use name="SimDataFormats/GeneratorProducts"/>
+<use name="FWCore/SharedMemory"/>
+<use name="clhep"/>
+<library name="GeneratorInterfaceCore_plugins" file="*.cc">
+  <flags EDM_PLUGIN="1"/>
+</library>
