+++ GeneratorInterface/Core/test/BuildFile.xml
@@ -1,30 +1,8 @@
 <environment>
-  <library file="FailingGeneratorFilter.cc" name="FailingGeneratorFilter">
-    <flags EDM_PLUGIN="1"/>
-    <use name="FWCore/Framework"/>
-    <use name="GeneratorInterface/Core"/>
-  </library>
-  <test name="FailingGeneratorFilterTest" command="${LOCALTOP}/src/GeneratorInterface/Core/test/test_FailingGeneratorFilter.sh"/>
-
-  <library file="TestInterProcessRandomProd.cc,RandomIntProducer.cc" name="TestInterProcessRandomProd">
-    <flags EDM_PLUGIN="1"/>
-    <use name="FWCore/AbstractServices"/>
-    <use name="FWCore/Framework"/>
-    <use name="FWCore/ParameterSet"/>
-    <use name="FWCore/Sources"/>
-    <use name="FWCore/SharedMemory"/>
-    <use name="FWCore/Utilities"/>
-    <use name="DataFormats/TestObjects"/>
-    <use name="GeneratorInterface/Core"/>
-    <use name="boost"/>
-    <use name="clhep"/>
-  </library>
-
-  <bin file="RandomIntProducer_t.cpp">
-    <use name="FWCore/Framework"/>
-    <use name="FWCore/ParameterSet"/>
-    <use name="FWCore/TestProcessor"/>
-    <use name="DataFormats/Provenance"/>
-    <use name="catch2"/>
-  </bin>
+<library file="FailingGeneratorFilter.cc" name="FailingGeneratorFilter">
+  <flags EDM_PLUGIN="1"/>
+  <use name="FWCore/Framework"/>
+  <use name="GeneratorInterface/Core"/>
+</library>
+<test name="FailingGeneratorFilterTest" command="${LOCALTOP}/src/GeneratorInterface/Core/test/test_FailingGeneratorFilter.sh"/>
 </environment>
